backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==4.2.0",
    "fastapi>=0.128.4",
    "greenlet>=3.3.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-chess>=1.999",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "stockfish>=4.0.6",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
